fix: address browser MCP review feedback
