+++ web/messages/es.json
@@ -870,6 +870,7 @@
       "tabsAndLogs": "Pestañas y registros",
       "targetingSurface": "Seleccionar una superficie de navegador",
       "targetingDesc": "La mayoría de los subcomandos requieren una superficie objetivo. Puede pasarla posicionalmente o con --surface.",
+      "mcpRegisterComment": "# Registra las herramientas MCP del navegador de cmux una vez",
       "navigation": "Navegación",
       "waitingSection": "Espera",
       "waitingDesc": "Use wait para bloquear hasta que se satisfagan selectores, texto, fragmentos de URL, estado de carga o una condición JavaScript.",
