+++ web/messages/it.json
@@ -870,6 +870,7 @@
       "tabsAndLogs": "Tab e log",
       "targetingSurface": "Targeting di una superficie browser",
       "targetingDesc": "La maggior parte dei sottocomandi richiede una superficie target. Puoi passarla posizionalmente o con --surface.",
+      "mcpRegisterComment": "# Registra una sola volta gli strumenti MCP del browser cmux",
       "navigation": "Navigazione",
       "waitingSection": "Attesa",
       "waitingDesc": "Usa wait per bloccare fino a quando selettori, testo, frammenti URL, stato di caricamento o una condizione JavaScript sono soddisfatti.",
