+++ web/messages/pt-BR.json
@@ -870,6 +870,7 @@
       "tabsAndLogs": "Abas e logs",
       "targetingSurface": "Direcionando uma superfície de navegador",
       "targetingDesc": "A maioria dos subcomandos requer uma superfície alvo. Você pode passá-la posicionalmente ou com --surface.",
+      "mcpRegisterComment": "# Registre as ferramentas MCP do navegador cmux uma vez",
       "navigation": "Navegação",
       "waitingSection": "Espera",
       "waitingDesc": "Use wait para bloquear até que seletores, texto, fragmentos de URL, estado de carregamento ou uma condição JavaScript seja satisfeita.",
