+++ web/messages/ru.json
@@ -870,6 +870,7 @@
       "tabsAndLogs": "Вкладки и логи",
       "targetingSurface": "Указание поверхности браузера",
       "targetingDesc": "Большинство подкоманд требуют целевую поверхность. Можно передать позиционно или через --surface.",
+      "mcpRegisterComment": "# Зарегистрируйте инструменты MCP браузера cmux один раз",
       "navigation": "Навигация",
       "waitingSection": "Ожидание",
       "waitingDesc": "Используй wait для блокировки до выполнения условия: селекторы, текст, фрагменты URL, состояние загрузки или JavaScript-условие.",
