+++ web/messages/tr.json
@@ -870,6 +870,7 @@
       "tabsAndLogs": "Sekmeler ve günlükler",
       "targetingSurface": "Tarayıcı yüzeyini hedefleme",
       "targetingDesc": "Çoğu alt komut bir hedef yüzey gerektirir. Konumsal olarak veya --surface ile geçirebilirsiniz.",
+      "mcpRegisterComment": "# cmux tarayıcı MCP araçlarını bir kez kaydet",
       "navigation": "Gezinme",
       "waitingSection": "Bekleme",
       "waitingDesc": "Seçiciler, metin, URL parçaları, yükleme durumu veya bir JavaScript koşulu karşılanana kadar engellemek için wait kullanın.",
