+++ web/messages/uk.json
@@ -871,6 +871,7 @@
       "tabsAndLogs": "Вкладки та журнали",
       "targetingSurface": "Вказання поверхні браузера",
       "targetingDesc": "Більшість підкоманд вимагають вказання цільової поверхні. Її можна передати позиційно або через --surface.",
+      "mcpRegisterComment": "# Зареєструйте інструменти MCP браузера cmux один раз",
       "navigation": "Навігація",
       "waitingSection": "Очікування",
       "waitingDesc": "Використовуйте wait для блокування до задоволення умов: селектори, текст, фрагменти URL, стан завантаження або умова JavaScript.",
