+++ web/messages/zh-CN.json
@@ -870,6 +870,7 @@
       "tabsAndLogs": "标签页和日志",
       "targetingSurface": "指定浏览器 surface",
       "targetingDesc": "大多数子命令需要指定目标 surface。可以通过位置参数或 --surface 传入。",
+      "mcpRegisterComment": "# 只需注册一次 cmux 浏览器 MCP 工具",
       "navigation": "导航",
       "waitingSection": "等待",
       "waitingDesc": "使用 wait 命令阻塞等待，直到选择器、文本、URL 片段、加载状态或 JavaScript 条件满足。",
